ms_ai/requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.35.0
python-dotenv==1.0.1
langchain-openai==0.3.7
langchain-core==0.3.40
langgraph>=0.0.40
chromadb==1.0.15
pydantic==2.7.4
openai==1.59.6
redis==5.0.8
websockets==13.1